+++ requirements.py
@@ -1,2 +1,3 @@
 numpy
 pytest
+pytype
