+++ setup.cfg
@@ -1,29 +1,31 @@
 [metadata]
-name = quantipy
+name = labelshift
 version = 0.0.1
 author = Paweł Czyż
 author_email = pczyz@protonmail.com
 description = 
 long_description = file: README.md
 long_description_content_type = text/markdown
-url = https://github.com/pawel-czyz/quantipy
+url = https://github.com/labelshift/labelshift
 project_urls =
-    Bug Tracker = https://github.com/pawel-czyz/quantipy/issues
+    Bug Tracker = https://github.com/labelshift/labelshift/issues
 classifiers =
     Programming Language :: Python :: 3
     License :: BSD-3-Clause
     Operating System :: OS Independent
 
+
 [options]
 package_dir =
-    = quantipy
+    = labelshift
 packages = find:
 python_requires = >=3.8
 install_requires =
     numpy >= 1.12,<2
 
+
 [options.packages.find]
-where = quantipy
+where = labelshift
 
 
 [yapf]
@@ -32,3 +34,7 @@ column_limit = 80
 spaces_before_comment = 2
 split_before_logical_operator = true
 split_before_arithmetic_operator = true
+
+
+[pytype]
+inputs = labelshift
